backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
google-adk
